Класс Index вынесен в отдельные файлы
